+++ src/components/common/ui/NotiBell.tsx
@@ -14,15 +14,24 @@ import { useRecoilState } from 'recoil';
 import useAlertModal from '../modal/AlertModal';
 import Image from 'next/image';
 import { userState } from '@/recoil/user';
+import useNotificationActions from '@/hooks/useNotificationActions';
 
 function NotiBell() {
   const { openModal } = useAlertModal();
   const [currentUser, setCurrentUser] = useRecoilState(userState);
   const [notificationData, setNotificationData] = useState<FlirtingListInNotificationType[]>([]);
   const [userNames, setUserNames] = useState<{ sender: string | null; receiver: string | null }[]>([]);
+
   const [filteredNotificationsSender, setFilteredNotificationsSender] = useState<FlirtingListInNotificationType[]>([]);
   const [filteredNotificationsReceiver, setFilteredNotificationsReceiver] = useState<FlirtingListInNotificationType[]>(
     []
+  );
+  const { popSenderData, popReceiverData, addReceiverData, addSenderData } = useNotificationActions(
+    notificationData,
+    filteredNotificationsSender,
+    filteredNotificationsReceiver,
+    setFilteredNotificationsSender,
+    setFilteredNotificationsReceiver
   );
 
   const fetchNotificationData = async () => {
@@ -92,32 +101,32 @@ function NotiBell() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [notificationData]);
 
-  const popSenderData = () => {
-    if (filteredNotificationsSender.length > 0 && filteredNotificationsSender[0].sender_is_read_in_noti === true) {
-      setFilteredNotificationsSender((prev) => prev.slice(1));
-    }
-  };
+  // const popSenderData = () => {
+  //   if (filteredNotificationsSender.length > 0 && filteredNotificationsSender[0].sender_is_read_in_noti === true) {
+  //     setFilteredNotificationsSender((prev) => prev.slice(1));
+  //   }
+  // };
 
-  const popReceiverData = () => {
-    if (
-      filteredNotificationsReceiver.length > 0 &&
-      filteredNotificationsReceiver[0].receiver_is_read_in_noti === true
-    ) {
-      setFilteredNotificationsReceiver((prev) => prev.slice(1));
-    }
-  };
+  // const popReceiverData = () => {
+  //   if (
+  //     filteredNotificationsReceiver.length > 0 &&
+  //     filteredNotificationsReceiver[0].receiver_is_read_in_noti === true
+  //   ) {
+  //     setFilteredNotificationsReceiver((prev) => prev.slice(1));
+  //   }
+  // };
 
-  const addReceiverData = () => {
-    if (notificationData.length > 0 && !notificationData[0].receiver_is_read_in_noti) {
-      setFilteredNotificationsReceiver((prev) => [...prev, notificationData[0]]);
-    }
-  };
+  // const addReceiverData = () => {
+  //   if (notificationData.length > 0 && !notificationData[0].receiver_is_read_in_noti) {
+  //     setFilteredNotificationsReceiver((prev) => [...prev, notificationData[0]]);
+  //   }
+  // };
 
-  const addSenderData = () => {
-    if (notificationData.length > 0 && !notificationData[0].sender_is_read_in_noti) {
-      setFilteredNotificationsSender((prev) => [...prev, notificationData[0]]);
-    }
-  };
+  // const addSenderData = () => {
+  //   if (notificationData.length > 0 && !notificationData[0].sender_is_read_in_noti) {
+  //     setFilteredNotificationsSender((prev) => [...prev, notificationData[0]]);
+  //   }
+  // };
 
   // console.log('filteredNotificationsSender', filteredNotificationsSender);
   // console.log('filteredNotificationsReceiver', filteredNotificationsReceiver);
